requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
streamlit
python-dotenv
serpapi
# plus any agent-specific deps like playwright or fpdf if you use them:
# playwright
# fpdf
